requirements.txt
pillow
colored
